feat: rename watchIntegration to watchDirectory
