+++ package/src/utilities/index.ts
@@ -4,6 +4,6 @@ export { addVitePlugin } from "./add-vite-plugin.js";
 export { hasIntegration } from "./has-integration.js";
 export { hasVitePlugin } from "./has-vite-plugin.js";
 export { injectDevRoute } from "./inject-dev-route.js";
-export { watchIntegration } from "./watch-integration.js";
+export { watchDirectory as watchIntegration } from "./watch-directory.js";
 export { addDevToolbarFrameworkApp } from "./add-devtoolbar-framework-app.js";
 export { addIntegration } from "./add-integration.js";
