+++ package/tests/unit/watch-integration.spec.ts
@@ -12,7 +12,7 @@ import {
 } from "vitest";
 import { createResolver } from "../../src/core/create-resolver.js";
 import type { HookParameters } from "../../src/core/types.js";
-import { watchIntegration } from "../../src/utilities/watch-integration.js";
+import { watchDirectory } from "../../src/utilities/watch-directory.ts";
 
 const tempFolderName = ".TMP_WATCHINTEGRATION";
 
@@ -69,14 +69,14 @@ describe("watchIntegration", () => {
 		const params = getParams();
 
 		expect(() =>
-			watchIntegration(params, resolve(tempFolderName)),
+			watchDirectory(params, resolve(tempFolderName)),
 		).not.toThrow();
 	});
 
 	test("Should call updateConfig", () => {
 		const params = getParams();
 
-		watchIntegration(params, resolve(tempFolderName));
+		watchDirectory(params, resolve(tempFolderName));
 
 		expect(params.updateConfig).toBeCalled();
 	});
@@ -84,7 +84,7 @@ describe("watchIntegration", () => {
 	test("Should call updateConfig once", () => {
 		const params = getParams();
 
-		watchIntegration(params, resolve(tempFolderName));
+		watchDirectory(params, resolve(tempFolderName));
 
 		expect(params.updateConfig).toBeCalledTimes(1);
 	});
@@ -92,7 +92,7 @@ describe("watchIntegration", () => {
 	test("Should call addWatchFile", () => {
 		const params = getParams();
 
-		watchIntegration(params, resolve(tempFolderName));
+		watchDirectory(params, resolve(tempFolderName));
 
 		expect(params.addWatchFile).toBeCalled();
 	});
@@ -100,7 +100,7 @@ describe("watchIntegration", () => {
 	test("Should call addWatchFile for each path (count)", () => {
 		const params = getParams();
 
-		watchIntegration(params, resolve(tempFolderName));
+		watchDirectory(params, resolve(tempFolderName));
 
 		const calls = params.addWatchFile.mock.calls.flatMap((entry) => entry[0]);
 
@@ -110,7 +110,7 @@ describe("watchIntegration", () => {
 	test("Should call addWatchFile for each path (path check)", () => {
 		const params = getParams();
 
-		watchIntegration(params, resolve(tempFolderName));
+		watchDirectory(params, resolve(tempFolderName));
 
 		const calls = params.addWatchFile.mock.calls.flatMap((entry) =>
 			normalize(entry[0].toString()),
@@ -134,7 +134,7 @@ describe("watchIntegration", () => {
 			}),
 		} as unknown as HookParameters<"astro:config:setup">;
 
-		watchIntegration(params, resolve(tempFolderName));
+		watchDirectory(params, resolve(tempFolderName));
 
 		// @ts-ignore - TS can't figure out that plugin _will_ actually be defined here
 		expect(plugin).toBeDefined();
@@ -149,7 +149,7 @@ describe("watchIntegration", () => {
 			}),
 		} as unknown as HookParameters<"astro:config:setup">;
 
-		watchIntegration(params, resolve(tempFolderName));
+		watchDirectory(params, resolve(tempFolderName));
 
 		// @ts-ignore - TS can't figure out that plugin _will_ actually be defined here
 		expect(plugin.name).toBeDefined();
